refactor(theme): rename internal token folders
