+++ ui-admin/theme/src/theme.css.ts
@@ -3,9 +3,9 @@ import './global.css'
 import { createTheme }         from '@vanilla-extract/css'
 import { createThemeContract } from '@vanilla-extract/css'
 
-import { darkThemeTokens }     from './theme/index.js'
-import { themeContract }       from './theme/index.js'
-import { lightThemeTokens }    from './theme/index.js'
+import { darkThemeTokens }     from './tokens/index.js'
+import { themeContract }       from './tokens/index.js'
+import { lightThemeTokens }    from './tokens/index.js'
 
 export const vars = createThemeContract(themeContract)
