+++ ui-parts/theme/src/theme.css.ts
@@ -2,15 +2,15 @@ import './global.css'
 
 import { createGlobalTheme } from '@vanilla-extract/css'
 
-import { borders }           from './theme/index.js'
-import { colors }            from './theme/index.js'
-import { fonts }             from './theme/index.js'
-import { fontSizes }         from './theme/index.js'
-import { fontWeights }       from './theme/index.js'
-import { lineHeights }       from './theme/index.js'
-import { radii }             from './theme/index.js'
-import { shadows }           from './theme/index.js'
-import { space }             from './theme/index.js'
+import { borders }           from './tokens/index.js'
+import { colors }            from './tokens/index.js'
+import { fonts }             from './tokens/index.js'
+import { fontSizes }         from './tokens/index.js'
+import { fontWeights }       from './tokens/index.js'
+import { lineHeights }       from './tokens/index.js'
+import { radii }             from './tokens/index.js'
+import { shadows }           from './tokens/index.js'
+import { space }             from './tokens/index.js'
 
 export const globalThemeId = 'atls-ui-parts-theme'
